test: rename all-FS test files to .int.test.ts
