+++ src/resource-diagnostics.int.test.ts
